Inverse Bedmap2 at 2km corrected version uoloaded
The inversion at 2km using Bedmap2 has been redone. This version has corrected the bug in the posFraction routine that led to values outside of the ice mask being used to calculate grounded fraction on u and v grids. This version also gives a standard value of WeertmanC (10^4) under ice shelves. This will be the value of WeertmanC under newly grounded ice in forward runs.