Inverse Bedmachine 3km corrected version uploaded
The inversion at 3km using BedmachineV2 has been redone. This version has corrected the bug in the posFraction routine that led to values outside of the ice mask being used to calculate grounded fraction on u and v grids. An error in the rockmask in the initial state has also been corrected.  This version also gives a standard value of WeertmanC (10^4) under ice shelves. This will be the value of WeertmanC under newly grounded ice in forward runs.