Add files via upload
BedmachineV3 at 2km ("full") initial files added for a version of the Matlab initialisation that has been edited to deal with the "stripy" velocities. In the new version the discretization of the shear stresses is done in a different way, which gets rid of the stripy effects from the velocities in the initial state, with a similar JRMS error as before (49 m/a after 20 iterations).